requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-spectacular==0.27.2
django-cors-headers==4.4.0

PyMySQL==1.1.1
django-environ==0.11.2
django-apscheduler==0.6.2
openpyxl==3.1.5
reportlab==4.2.2
